+++ Libraries/MiKTeX/Setup/SetupService.cpp
@@ -88,7 +88,7 @@ string IssueSeverityString(IssueSeverity severity)
   case IssueSeverity::Critical: return "critical issue";
   case IssueSeverity::Major: return "major issue";
   case IssueSeverity::Minor: return "minor issue";
-  case IssueSeverity::Trivial: return "trivial issue";
+  case IssueSeverity::Trivial: return "slight issue";
   default: MIKTEX_UNEXPECTED();
   }
 }
@@ -1973,6 +1973,28 @@ vector<Issue> SetupServiceImpl::FindIssues(bool checkPath, bool checkPackageInte
       });
     }
   }
+#if defined(MIKTEX_WINDOWS_32)
+  IssueSeverity s = IssueSeverity::Trivial;
+  if (now > 1760443200)
+  {
+    s = IssueSeverity::Critical;
+  }
+  else if (now > 1730808000)
+  {
+    s = IssueSeverity::Major;
+  }
+  else if (now > 1667908800)
+  {
+    s = IssueSeverity::Minor;
+  }
+  result.push_back({
+    IssueType::Windows32bit,
+    s,
+    T_("You are running the unsupported 32-bit version of MiKTeX for Windows."),
+    T_("Install the 64-bit version of MiKTeX for Windows."),
+    "miktex-32bit"
+  });
+#endif
   if (session->IsSharedSetup())
   {
     InstallationSummary commonInstallation = packageManager->GetInstallationSummary(false);